pyproject.toml
[tool.poetry]
name = "myfood-compute"
version = "0.1.0"
description = ""
authors = ["TY <tyyyy66@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
fastapi-cli = "^0.0.7"
requests = "^2.32.3"
opencv-python = "^4.11.0.86"
pyqt5 = "^5.15.11"
torch = "2.2.2"
ultralytics = "^8.3.102"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.20"
numpy = "^1.24"
pydantic-settings = "^2.8.1"
asyncpg = "^0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
